use native Device (#7511)
