+++ cocos/core/3d/builtin/init.ts
@@ -3,7 +3,7 @@ import { ImageAsset } from '../../assets/image-asset';
 import { SpriteFrame } from '../../assets/sprite-frame';
 import { Texture2D } from '../../assets/texture-2d';
 import { TextureCube } from '../../assets/texture-cube';
-import { GFXDevice } from '../../gfx/device';
+import { GFXDevice } from '../../gfx';
 import effects from './effects';
 import { legacyCC } from '../../global-exports';
