+++ cocos/core/assets/texture-base.ts
@@ -31,12 +31,11 @@
 
 // @ts-check
 import {ccclass, serializable} from 'cc.decorator';
-import { GFXDevice } from '../gfx/device';
 import { genSamplerHash, SamplerInfoIndex, samplerLib } from '../renderer/core/sampler-lib';
 import IDGenerator from '../utils/id-generator';
 import { Asset } from './asset';
 import { Filter, PixelFormat, WrapMode } from './asset-enum';
-import { GFXSampler, GFXTexture } from '../gfx';
+import { GFXSampler, GFXTexture, GFXDevice } from '../gfx';
 import { legacyCC } from '../global-exports';
 import { errorID } from '../platform/debug';
