+++ cocos/core/pipeline/render-batched-queue.ts
@@ -6,8 +6,7 @@
 import { GFXCommandBuffer } from '../gfx/command-buffer';
 import { BatchedBuffer } from './batched-buffer';
 import { PipelineStateManager } from './pipeline-state-manager';
-import { GFXDevice } from '../gfx/device';
-import { GFXRenderPass } from '../gfx';
+import { GFXRenderPass, GFXDevice } from '../gfx';
 import { DSPool, ShaderPool, PassPool, PassView } from '../renderer/core/memory-pools';
 import { SetIndex } from './define';
