+++ cocos/core/renderer/core/pass.ts
@@ -34,7 +34,6 @@ import { IPassInfo, IPassStates, IPropertyInfo } from '../../assets/effect-asset
 import { TextureBase } from '../../assets/texture-base';
 import { GFXDescriptorSet, GFXDescriptorSetInfo } from '../../gfx/descriptor-set';
 import { GFXBuffer, GFXBufferInfo, GFXBufferViewInfo } from '../../gfx/buffer';
-import { GFXFeature, GFXDevice } from '../../gfx/device';
 import { GFXBlendState, GFXBlendTarget, GFXDepthStencilState, GFXRasterizerState } from '../../gfx/pipeline-state';
 import { GFXSampler } from '../../gfx/sampler';
 import { RenderPassStage, RenderPriority, SetIndex } from '../../pipeline/define';
@@ -48,8 +47,8 @@ import { PassView, BlendStatePool, RasterizerStatePool, DepthStencilStatePool,
 import { customizeType, getBindingFromHandle, getPropertyTypeFromHandle, getDefaultFromType,
     getOffsetFromHandle, getTypeFromHandle, MacroRecord, MaterialProperty, type2reader, type2writer, PropertyType } from './pass-utils';
 import { GFXBufferUsageBit, GFXGetTypeSize, GFXMemoryUsageBit, GFXPrimitiveMode,
-    GFXType, GFXDynamicStateFlagBit, GFXDynamicStateFlags } from '../../gfx/define';
-import { GFXDescriptorSetLayoutInfo, GFXTexture } from '../../gfx';
+    GFXType, GFXDynamicStateFlagBit, GFXDynamicStateFlags, GFXFeature } from '../../gfx/define';
+import { GFXDescriptorSetLayoutInfo, GFXTexture,  GFXDevice } from '../../gfx';
 
 export interface IPassInfoFull extends IPassInfo {
     // generated part
