+++ cocos/core/renderer/core/program-lib.ts
@@ -29,8 +29,7 @@
  */
 
 import { IBlockInfo, IBuiltinInfo, IDefineInfo, IShaderInfo } from '../../assets/effect-asset';
-import { GFXDescriptorType, GFXGetTypeSize, GFXShaderStageFlagBit } from '../../gfx/define';
-import { GFXAPI, GFXDevice } from '../../gfx/device';
+import { GFXDescriptorType, GFXGetTypeSize, GFXShaderStageFlagBit, GFXAPI } from '../../gfx/define';
 import { GFXAttribute } from '../../gfx/input-assembler';
 import { GFXUniformBlock, GFXShaderInfo, GFXUniformSampler, GFXUniform, GFXShaderStage } from '../../gfx/shader';
 import { SetIndex, IDescriptorSetLayoutInfo, globalDescriptorSetLayout, localDescriptorSetLayout } from '../../pipeline/define';
@@ -40,7 +39,7 @@ import { legacyCC } from '../../global-exports';
 import { ShaderPool, ShaderHandle, PipelineLayoutHandle, PipelineLayoutPool, NULL_HANDLE } from './memory-pools';
 import { DESCRIPTOR_SAMPLER_TYPE, DESCRIPTOR_BUFFER_TYPE } from '../../gfx/descriptor-set';
 import { GFXDescriptorSetLayout, GFXDescriptorSetLayoutBinding, GFXDescriptorSetLayoutInfo } from '../../gfx/descriptor-set-layout';
-import { GFXPipelineLayoutInfo } from '../../gfx';
+import { GFXPipelineLayoutInfo, GFXDevice } from '../../gfx';
 
 const _dsLayoutInfo = new GFXDescriptorSetLayoutInfo();
