+++ cocos/core/renderer/models/skeletal-animation-utils.ts
@@ -36,8 +36,8 @@ import { Mesh } from '../../assets/mesh';
 import { Skeleton } from '../../assets/skeleton';
 import { aabb } from '../../geometry';
 import { GFXBuffer, GFXBufferInfo } from '../../gfx/buffer';
-import { GFXAddress, GFXBufferUsageBit, GFXFilter, GFXFormat, GFXFormatInfos, GFXMemoryUsageBit } from '../../gfx/define';
-import { GFXDevice, GFXFeature } from '../../gfx/device';
+import { GFXAddress, GFXBufferUsageBit, GFXFilter, GFXFormat, GFXFormatInfos, GFXMemoryUsageBit, GFXFeature } from '../../gfx/define';
+import { GFXDevice } from '../../gfx';
 import { Mat4, Quat, Vec3 } from '../../math';
 import { UBOSkinningAnimation } from '../../pipeline/define';
 import { Node } from '../../scene-graph';
