+++ cocos/core/renderer/scene/submodel.ts
@@ -1,10 +1,9 @@
 import { RenderingSubMesh } from '../../assets/mesh';
-import { GFXDevice } from '../../gfx/device';
 import { GFXInputAssembler } from '../../gfx/input-assembler';
 import { RenderPriority, SetIndex } from '../../pipeline/define';
 import { IMacroPatch, Pass } from '../core/pass';
 import { DSPool, IAPool, SubModelPool, SubModelView, SubModelHandle, NULL_HANDLE } from '../core/memory-pools';
-import { GFXDescriptorSet, GFXDescriptorSetInfo } from '../../gfx';
+import { GFXDescriptorSet, GFXDescriptorSetInfo, GFXDevice } from '../../gfx';
 import { legacyCC } from '../../global-exports';
 
 const _dsInfo = new GFXDescriptorSetInfo(null!);
