+++ cocos/core/renderer/ui/ui.ts
@@ -30,10 +30,9 @@
 import { UIStaticBatch } from '../../../ui';
 import { Material } from '../../assets/material';
 import { Canvas, UIComponent, UIRenderable } from '../../components/ui-base';
-import { GFXDevice } from '../../gfx/device';
 import { GFXAttribute } from '../../gfx/input-assembler';
 import { GFXSampler } from '../../gfx/sampler';
-import { GFXTexture } from '../../gfx';
+import { GFXTexture, GFXDevice } from '../../gfx';
 import { Pool, RecyclePool } from '../../memop';
 import { CachedArray } from '../../memop/cached-array';
 import { Camera } from '../scene/camera';
