+++ cocos/core/splash-screen.ts
@@ -7,13 +7,12 @@ import * as easing from './animation/easing';
 import { Material } from './assets/material';
 import { GFXBuffer, GFXBufferInfo } from './gfx/buffer';
 import { GFXCommandBuffer } from './gfx/command-buffer';
-import { GFXDevice } from './gfx/device';
 import { GFXFramebuffer } from './gfx/framebuffer';
 import { GFXInputAssembler, GFXInputAssemblerInfo, GFXAttribute } from './gfx/input-assembler';
 import { clamp01 } from './math/utils';
 import { COCOSPLAY, XIAOMI, JSB } from 'internal:constants';
 import { sys } from './platform/sys';
-import { GFXSampler, GFXSamplerInfo, GFXShader, GFXTexture, GFXTextureInfo } from './gfx';
+import { GFXSampler, GFXSamplerInfo, GFXShader, GFXTexture, GFXTextureInfo, GFXDevice } from './gfx';
 import { PipelineStateManager } from './pipeline';
 import { legacyCC } from './global-exports';
 import { Root } from './root';
