+++ cocos/core/utils/profiler/profiler.ts
@@ -29,8 +29,7 @@ import { Camera } from '../../3d/framework/camera-component';
 import { createMesh } from '../../3d/misc/utils';
 import { Material } from '../../assets/material';
 import { GFXBufferTextureCopy, GFXClearFlag, GFXFormat, GFXTextureType, GFXTextureUsageBit } from '../../gfx/define';
-import { GFXDevice } from '../../gfx/device';
-import { GFXTexture, GFXTextureInfo } from '../../gfx';
+import { GFXTexture, GFXTextureInfo, GFXDevice } from '../../gfx';
 import { Vec4 } from '../../math';
 import { Layers } from '../../scene-graph';
 import { Node } from '../../scene-graph/node';
