+++ cocos/particle/renderer/trail.ts
@@ -10,7 +10,7 @@ import { ccclass, tooltip, displayOrder, type, serializable } from 'cc.decorator
 import { director } from '../../core/director';
 import { GFX_DRAW_INFO_SIZE, GFXBuffer, GFXIndirectBuffer, GFXBufferInfo, GFXDrawInfo } from '../../core/gfx/buffer';
 import { GFXAttributeName, GFXBufferUsageBit, GFXFormat, GFXFormatInfos, GFXMemoryUsageBit, GFXPrimitiveMode } from '../../core/gfx/define';
-import { GFXDevice } from '../../core/gfx/device';
+import { GFXDevice } from '../../core/gfx';
 import { GFXAttribute } from '../../core/gfx/input-assembler';
 import { Color, Mat4, Quat, toRadian, Vec3 } from '../../core/math';
 import { Pool } from '../../core/memop';
